backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
loguru
alembic
slowapi
openpyxl
httpx
pytest
